+++ jest.config.js
@@ -17,6 +17,7 @@ module.exports = {
     "**/__tests__/**/*.+(ts|tsx|js)",
     "**/?(*.)+(spec|test).+(ts|tsx|js)",
   ],
+  timers: 'fake',
   transform: {
     "^.+\\.(ts|tsx)$": "ts-jest",
   },