+++ src/components/NotificationQueue/index.tsx
@@ -5,9 +5,19 @@ import Style from './style';
 
 const NotificationQueue = (): JSX.Element => {
   const { state } = React.useContext(Context);
-  const didMount = React.useRef(false);
-  const [queue, setQueue] = React.useState([] as string[]);
   const { canDrawCard } = state;
+
+  const didMount = React.useRef(false);
+  const timerRef = React.useRef(0);
+
+  const [queue, setQueue] = React.useState([] as string[]);
+  const [timer, setTimer] = React.useState(Date.now);
+
+  const startTimer = () => {
+    timerRef.current = setTimeout(() => {
+      setTimer(Date.now);
+    }, 1500);
+  };
 
   React.useEffect(() => {
     if (canDrawCard && didMount.current) {
@@ -16,6 +26,15 @@ const NotificationQueue = (): JSX.Element => {
       didMount.current = true;
     }
   }, [canDrawCard]);
+
+  React.useEffect(() => {
+    if (timerRef.current === 0 && queue.length !== 0) startTimer();
+  }, [queue]);
+
+  React.useEffect(() => {
+    timerRef.current = 0;
+    setQueue(queue.slice(1));
+  }, [timer]);
 
   return (
     <Style className='notificationQueue'>